requirements.txt
beautifulsoup4==4.12.2
pandas==1.5.3
requests==2.31.0
python-slugify==8.0.1
streamlit==1.26.0
